services/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.7.1
kafka-python==2.0.2
